Split tag command
